Upgrade ERD to Class Diagram; fix iCare FK type; add MySQL setup
Summary:
Replaced the Mermaid erDiagram in README.md with a full UML classDiagram that maps each database table to its ORM class, carries SQL column types, shows key ORM methods, and uses proper UML relationship notation (composition, association, cardinality). The class diagram and ERD are now unified — every entity in the diagram corresponds 1-to-1 with a Propel model class and its backing table, making it the single canonical data-model reference. Fixed a type mismatch in 7.4.0-icare.sql where icare_attendance.person_id was declared INT UNSIGNED instead of MEDIUMINT(9) UNSIGNED, causing a foreign-key constraint error against person_per.per_ID on MySQL 8.4. Added docker/setup-local-mysql.sql as a one-shot helper for creating the local churchcrm database and user.

Details:
1. README.md: replaced erDiagram block with classDiagram covering Member, Family, IcareGroup, IcareMembership, IcareMeeting, IcareAttendance, IcareVisitor, Event, EventAttendance, EventCount, KioskDevice, User — each with SQL column types, PK/FK annotations, and ORM methods
2. README.md: Entity Overview table updated to show ORM class names alongside DB table names
3. 7.4.0-icare.sql: fixed icare_attendance.person_id from INT UNSIGNED to MEDIUMINT(9) UNSIGNED to match person_per.per_ID column type
4. docker/setup-local-mysql.sql: new script — creates churchcrm database + churchcrm user (password: changeme); run once as root before applying Install.sql
5. Local MySQL 8.4.9 deployed: initialized at ~/MySQL/data, service running on localhost:3306, churchcrm schema (49 tables) applied and verified, all iCare FK constraints pass

+++ src/mysql/upgrade/7.4.0-icare.sql
@@ -16,9 +16,9 @@ CREATE TABLE `icare_meeting` (
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
 
 CREATE TABLE `icare_attendance` (
-    `id`          INT UNSIGNED NOT NULL AUTO_INCREMENT,
-    `meeting_id`  INT UNSIGNED NOT NULL,
-    `person_id`   INT UNSIGNED NOT NULL,
+    `id`          INT UNSIGNED          NOT NULL AUTO_INCREMENT,
+    `meeting_id`  INT UNSIGNED          NOT NULL,
+    `person_id`   MEDIUMINT(9) UNSIGNED NOT NULL,
     `recorded_at` TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (`id`),
     UNIQUE KEY `icare_attend_unique` (`meeting_id`, `person_id`),
